@@ -5,29 +5,40 @@
   "homepage": "https://caguy.github.io/training-assistant",
   "keywords": [],
   "main": "src/index.js",
+  "type": "module",
   "dependencies": {
-    "@emotion/react": "11.1.4",
-    "@emotion/styled": "11.0.0",
+    "@emotion/react": "11.11.1",
+    "@emotion/styled": "11.11.0",
     "emotion-normalize": "11.0.1",
-    "lodash": "4.17.20",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-flip-toolkit": "^7.0.13",
-    "react-redux": "7.2.2",
-    "react-scripts": "4.0.2",
-    "redux": "4.0.5",
-    "redux-thunk": "2.3.0"
+    "lodash": "4.17.21",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-flip-toolkit": "^7.1.0",
+    "react-redux": "9.0.2",
+    "redux": "5.0.0",
+    "redux-thunk": "3.1.0"
   },
   "devDependencies": {
-    "gh-pages": "^3.1.0",
-    "prop-types": "15.7.2",
-    "typescript": "3.3.3"
+    "@types/node": "^20.10.4",
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "eslint": "^8.53.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.4",
+    "gh-pages": "^6.1.0",
+    "prop-types": "15.8.1",
+    "typescript": "^5.2.2",
+    "vite": "^5.0.0",
+    "vite-plugin-svgr": "^4.2.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build"
   },
