@@ -1,20 +1,23 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 import { ThemeProvider, Global, css } from "@emotion/react";
 import emotionNormalize from "emotion-normalize";
 
-import store from "./state/store";
-import { StatConverter } from "./components";
-import { theme, globalStyles } from "./styles";
+import { store } from "@/state/store";
+import { StatConverter } from "@/components";
+import { theme, globalStyles } from "@/styles";
 
 const rootElement = document.getElementById("root");
-ReactDOM.render(
+ReactDOM.createRoot(rootElement).render(
   <Provider store={store}>
-    <Global styles={css`${globalStyles}${emotionNormalize}`} />
+    <Global
+      styles={css`
+        ${globalStyles}${emotionNormalize}
+      `}
+    />
     <ThemeProvider theme={theme}>
       <StatConverter />
     </ThemeProvider>
-  </Provider>,
-  rootElement
+  </Provider>
 );
