@@ -1,17 +1,17 @@
-import { SPEED, PACE, TIME, DISTANCE } from "state/constants";
-import { SET_TOTAL } from "state/actionTypes";
+import { SPEED, PACE, TIME, DISTANCE } from "@/state/constants";
+import { SET_TOTAL } from "@/state/actionTypes";
 
 const defaultState = {
   [SPEED.type]: {
     value: 10,
-    unit: SPEED.unit
+    unit: SPEED.unit,
   },
   [PACE.type]: {
     value: { min: 6, sec: 0 },
-    unit: PACE.unit
+    unit: PACE.unit,
   },
   [TIME.type]: { value: { h: 1, min: 0, sec: 0 }, unit: TIME.unit },
-  [DISTANCE.type]: { value: 10, unit: DISTANCE.unit }
+  [DISTANCE.type]: { value: 10, unit: DISTANCE.unit },
 };
 
 export default function total(state = defaultState, action) {
